+++ src/os/mod.rs
@@ -15,6 +15,8 @@ pub mod macos;
 pub mod unix;
 pub mod windows;
 
+use std::os::raw;
+
 /// Platform-specific extensions for OpenGL contexts.
 pub trait GlContextExt {
     /// Raw context handle.
@@ -22,4 +24,11 @@ pub trait GlContextExt {
 
     /// Returns the raw context handle.
     unsafe fn raw_handle(&self) -> Self::Handle;
+
+    /// Returns a pointer to the `EGLDisplay` object of EGL that is used by this context.
+    ///
+    /// Return `None` if the context doesn't use EGL.
+    //
+    // The pointer will become invalid when the glutin `GlContext` is destroyed.
+    unsafe fn get_egl_display(&self) -> Option<*const raw::c_void>;
 }